make hull int value

+++ src/main/java/it/spaghettisource/navaltrader/game/model/Company.java
@@ -78,7 +78,7 @@ public class Company implements Entity {
 		
 	}
 	
-	public void repairShip(String shipName,Double amountToRepair,Double priceToPay) {
+	public void repairShip(String shipName,int amountToRepair,Double priceToPay) {
 		Ship ship = getShipByName(shipName);
 		ship.addHull(amountToRepair);
 		removeBudget(priceToPay);
